backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.2
PyJWT==2.10.1
sqlparse==0.5.5
typing_extensions==4.15.0
